+++ R/param.R
@@ -293,6 +293,10 @@ ParametricJob <- R6::R6Class(classname = "ParametricJob", cloneable = FALSE,
         #'        `FALSE`.
         #' @param echo Only applicable when `wait` is `TRUE`. Whether to
         #'        simulation status. Default: same as `wait`.
+        #' @param separate If `TRUE`, all models are saved in a separate folder
+        #'        with each model's name under `dir` when simulation. If `FALSE`,
+        #'        all models are saved in `dir` when simulation. Default:
+        #'        `TRUE`.
         #'
         #' @return The `ParametricJob` object itself, invisibly.
         #'
@@ -307,8 +311,8 @@ ParametricJob <- R6::R6Class(classname = "ParametricJob", cloneable = FALSE,
         #' print(param)
         #' }
         #'
-        run = function (dir = NULL, wait = TRUE, force = FALSE, copy_external = FALSE, echo = wait)
-            param_run(self, private, dir, wait, force, copy_external, echo),
+        run = function (dir = NULL, wait = TRUE, force = FALSE, copy_external = FALSE, echo = wait, separate = TRUE)
+            param_run(self, private, dir, wait, force, copy_external, echo, separate),
         # }}}
 
         # print {{{
@@ -468,7 +472,9 @@ param_apply_measure <- function (self, private, measure, ..., .names = NULL, .en
 }
 # }}}
 # param_run {{{
-param_run <- function (self, private, output_dir = NULL, wait = TRUE, force = FALSE, copy_external = FALSE, echo = wait) {
+param_run <- function (self, private, output_dir = NULL, wait = TRUE,
+                       force = FALSE, copy_external = FALSE, echo = wait,
+                       separate = TRUE) {
     if (is.null(private$m_idfs)) {
         abort("No measure has been applied.")
     }
@@ -489,7 +495,7 @@ param_run <- function (self, private, output_dir = NULL, wait = TRUE, force = FA
     }
 
     private$log_new_uuid()
-    epgroup_run_models(self, private, output_dir, wait, force, copy_external, echo)
+    epgroup_run_models(self, private, output_dir, wait, force, copy_external, echo, separate)
 }
 # }}}
 # param_save {{{