+++ NEWS.md
@@ -7,6 +7,10 @@
 * Add EnergyPlus v9.3 and v9.4 support (#343, #347, #369).
 * Now `ParametricJob$apply_measure()` will give a nice progress bar when
   parametric models are creating (#378).
+* A new argument `separate` is added in `EplusGroupJob$run()` and
+  `ParametricJob$run()` with default value being `TRUE`. If set to `FALSE`, all
+  models are saved in `dir` when simulation, instead of creating a folder for
+  each model and running simulation there (#381).
 
 ## Major changes